requirements.txt
flask==2.2.3
pymongo==4.3.3
python-dotenv==1.0.0
openai==1.3.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
flask-cors==3.0.10
gunicorn==21.2.0
